Remove needless NULL checks
We should report some messages for the unexpected case
instead of just ignoring.

We should write a code for avoiding the unexpected case
rather than checking it. So we remove the checks. If we add
the checks again, we should also add an error or warning
message for fixing our code.