Unwind jump buffers to avoid an invalid memory access on an unhandled exception
If jump buffers are stacked, the previous catch handler calls
longjmp() with the top of the jump buffers. However, it is no
longer valid, because it was on a stack unwound by the exception.
This situation typically happens when an exception is thrown in
cutter's assertion function.

This patch forces to use the base jump buffer when an unhandled
exception is caught.